pyproject.toml
[tool.poetry]
name = "discoflix"
version = "0.1.0"
description = ""
authors = ["Nick Heyer <nick@heyer.app>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
channels = "4.1.0"
discord-py = "2.4.0"
django = "5.1.2"
drf-spectacular = "0.27.2"
python-dotenv = "1.0.1"
websockets = "13.1"
daphne = "^4.1.2"
django-cors-headers = "^4.6.0"
django-jazzmin = "^3.0.1"
requests = "^2.32.3"
whitenoise = "^6.8.2"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
psutil = "^6.1.0"
openai = "^1.55.2"


[tool.poetry.group.dev.dependencies]
iniconfig = "^2.0.0"
pluggy = "^1.5.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-django = "^4.9.0"
dpytest = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
